WIP: Flip the map_serializers dependency
