+++ src/ds/map_serializers.h
@@ -1,62 +1,15 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the Apache 2.0 License.
 #pragma once
+#include "ds/serialized.h"
 
-#include "ccf/byte_vector.h"
-#include "ccf/ccf_assert.h"
-#include "serialized.h"
-
+// TODO: This shouldn't be done like this
 namespace map
 {
-  using Version = uint64_t;
-  using DeletableVersion = int64_t;
-
-  template <typename V>
-  struct VersionV
-  {
-    DeletableVersion version;
-    Version read_version;
-    V value;
-
-    VersionV() :
-      version(std::numeric_limits<decltype(version)>::min()),
-      read_version(std::numeric_limits<decltype(read_version)>::min())
-    {}
-    VersionV(DeletableVersion ver, Version read_ver, V val) :
-      version(ver),
-      read_version(read_ver),
-      value(val)
-    {}
-  };
-
-  namespace serialisers
-  {
-    using SerialisedEntry = ccf::ByteVector;
-  }
-
-  namespace untyped
-  {
-    using SerialisedEntry = serialisers::SerialisedEntry;
-    using VersionV = VersionV<SerialisedEntry>;
-  }
-
   template <class T>
   inline size_t get_size(const T& data)
   {
     return sizeof(uint64_t) + sizeof(data);
-  }
-
-  template <>
-  inline size_t get_size<untyped::SerialisedEntry>(
-    const untyped::SerialisedEntry& data)
-  {
-    return sizeof(uint64_t) + data.size();
-  }
-
-  template <>
-  inline size_t get_size<untyped::VersionV>(const untyped::VersionV& data)
-  {
-    return sizeof(uint64_t) + sizeof(data.version) + data.value.size();
   }
 
   static uint32_t get_padding(uint32_t size)
@@ -81,12 +34,11 @@ namespace map
     return padding_size;
   }
 
-  template <class K, class V>
-  static size_t get_size_with_padding(const K& k, const V& v)
+  template <class T>
+  static size_t get_size_with_padding(const T& t)
   {
-    uint32_t size_k = get_size(k);
-    uint32_t size_v = get_size(v);
-    return size_k + get_padding(size_k) + size_v + get_padding(size_v);
+    const uint32_t t_size = get_size(t);
+    return t_size + get_padding(t_size);
   }
 
   template <class T>
@@ -103,44 +55,6 @@ namespace map
     return sizeof(uint64_t) + sizeof(T);
   }
 
-  template <>
-  inline size_t serialize<untyped::SerialisedEntry>(
-    const untyped::SerialisedEntry& t, uint8_t*& data, size_t& size)
-  {
-    uint64_t data_size = t.size();
-    serialized::write(
-      data,
-      size,
-      reinterpret_cast<const uint8_t*>(&data_size),
-      sizeof(uint64_t));
-    serialized::write(
-      data, size, reinterpret_cast<const uint8_t*>(t.data()), data_size);
-    return sizeof(uint64_t) + data_size;
-  }
-
-  template <>
-  inline size_t serialize<untyped::VersionV>(
-    const untyped::VersionV& t, uint8_t*& data, size_t& size)
-  {
-    uint64_t data_size = sizeof(t.version) + t.value.size();
-    serialized::write(
-      data,
-      size,
-      reinterpret_cast<const uint8_t*>(&data_size),
-      sizeof(uint64_t));
-    serialized::write(
-      data,
-      size,
-      reinterpret_cast<const uint8_t*>(&t.version),
-      sizeof(t.version));
-    serialized::write(
-      data,
-      size,
-      reinterpret_cast<const uint8_t*>(t.value.data()),
-      t.value.size());
-    return sizeof(uint64_t) + sizeof(t.version) + t.value.size();
-  }
-
   template <class T>
   inline T deserialize(const uint8_t*& data, size_t& size)
   {
@@ -149,31 +63,6 @@ namespace map
     CCF_ASSERT_FMT(
       result == sizeof(T), "result:{} == sizeof(T):{}", result, sizeof(T));
     return serialized::read<T>(data, size);
-  }
-
-  template <>
-  inline untyped::SerialisedEntry deserialize<untyped::SerialisedEntry>(
-    const uint8_t*& data, size_t& size)
-  {
-    uint64_t data_size = serialized::read<uint64_t>(data, size);
-    untyped::SerialisedEntry ret;
-    ret.append(data, data + data_size);
-    serialized::skip(data, size, data_size);
-    return ret;
-  }
-
-  template <>
-  inline untyped::VersionV deserialize<untyped::VersionV>(
-    const uint8_t*& data, size_t& size)
-  {
-    untyped::VersionV ret;
-    uint64_t data_size = serialized::read<uint64_t>(data, size);
-    Version version = serialized::read<Version>(data, size);
-    ret.version = version;
-    data_size -= sizeof(Version);
-    ret.value.append(data, data + data_size);
-    serialized::skip(data, size, data_size);
-    return ret;
   }
 
   template <class M>
